requirement.txt
python-dotenv==1.0.0
sentence-transformers==2.2.2
openai==1.53.0
faiss-cpu==1.7.4
numpy==1.23.5
PyMuPDF==1.22.5
spacy==3.5.3
slack_sdk==3.21.3